Addressed the review comments to use the functions in utils.py
