+++ smdebug/pytorch/utils.py
@@ -1,6 +1,7 @@
 # Third Party
 import numpy as np
 import torch
+from packaging import version
 
 # First Party
 from smdebug.core.reduction_config import ALLOWED_NORMS, ALLOWED_REDUCTIONS
@@ -34,3 +35,27 @@ def get_reduction_of_data(reduction_name, tensor_data, tensor_name, abs=False):
         op = f(tensor_data)
         return op
     raise RuntimeError("Invalid reduction_name {0}".format(reduction_name))
+
+
+def is_pt_1_5():
+    """
+    Determine whether the version of torch is 1.5.x
+    :return: bool
+    """
+    return version.parse("1.5.0") <= version.parse(torch.__version__) < version.parse("1.6.0")
+
+
+def is_pt_1_6():
+    """
+    Determine whether the version of torch is 1.6.x
+    :return: bool
+    """
+    return version.parse("1.6.0") <= version.parse(torch.__version__) < version.parse("1.7.0")
+
+
+def is_pt_1_7():
+    """
+    Determine whether the version of torch is 1.7.x
+    :return: bool
+    """
+    return version.parse("1.7.0") <= version.parse(torch.__version__) < version.parse("1.8.0")
